server/requirements.txt
absl-py==1.3.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autosemver==0.5.5
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.11.1
blis==0.7.9
Brotli==1.0.9
bs4==0.0.1
build==0.10.0
CacheControl==0.12.14
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
ci-info==0.3.0
cleantext==1.1.4
cleo==2.0.1
click==7.1.2
colorama==0.4.6
comm==0.1.3
confection==0.0.4
configobj==5.0.8
configparser==6.0.0
contourpy==1.0.7
crashtest==0.4.1
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dulwich==0.19.16
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl
en-use-md @ https://github.com/MartinoMensio/spacy-universal-sentence-encoder/releases/download/v0.4.3/en_use_md-0.4.3.tar.gz
etelemetry==0.3.0
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.2
flatbuffers==2.0.7
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
Fuzzy==1.2.2
gast==0.4.0
gensim==4.3.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-search-results==2.4.2
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==5.1.0
inspire-utils==3.0.36
installer==0.7.0
ipykernel==6.22.0
ipython==8.12.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.11.0
keyring==23.13.1
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
langid==1.1.6
Levenshtein==0.23.0
libclang==14.0.6
lockfile==0.12.2
looseversion==1.3.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
mplcursors==0.5.2
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
nameparser==0.5.8
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.7
numpy==1.24.2
oauthlib==3.2.2
onnx==1.12.0
openai==0.27.6
opencv-contrib-python==4.7.0.68
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==22.0
pandas==1.3.5
parsel==1.7.0
parso==0.8.3
pathlib==1.0.1
pathy==0.10.1
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.9.6
platformdirs==3.8.0
poetry-core==1.6.1
poetry-plugin-export==1.4.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.19.6
prov==2.0.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.7.4
pydot==1.4.2
pydyf==0.7.0
Pygments==2.15.0
PyMuPDF==1.22.5
pyparsing==3.0.9
PyPDF2==1.28.6
pypdfium2==3.21.1
pyphen==0.14.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-magic-bin==0.4.14
pytz==2022.6
pywin32==306
pywin32-ctypes==0.2.2
pyxnat==1.5
PyYAML==6.0
pyzmq==25.0.2
rapidfuzz==3.5.2
rdflib==6.3.2
refextract==1.1.4
regex==2022.10.31
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Shapely==1.8.5.post1
shellingham==1.5.0.post1
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.2
spacy-universal-sentence-encoder==0.4.5
srsly==2.4.5
stack-data==0.6.2
starlette==0.28.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.20.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
tf2onnx==1.14.0
thinc==8.1.8
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.8
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
traits==6.3.2
transformers==4.26.1
trove-classifiers==2023.5.24
typeguard==2.13.3
typer==0.3.2
typing_extensions==4.4.0
Unidecode==1.3.6
urllib3==1.26.12
uvicorn==0.22.0
virtualenv==20.23.1
w3lib==2.1.1
wasabi==0.10.1
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.2.2
word2vec==0.11.1
wrapt==1.14.1
yarl==1.9.2
zipp==3.11.0
zopfli==0.2.2
